Use dicts for output of distributedAeroLoads()
Makes it easier to add more fields and not have to worry about ordering.  Standardizes the output which helps get rid of weird logic checking. Also doesn't store anything to self.a and self.ap anymore - those weren't used.

I also tried this at the __loads() level, but because that is called in a loop across blade stations, the performance hit of building and extracting from all these dicts was too much.